+++ swift-sdk/Core/Constants.swift
@@ -77,6 +77,9 @@ enum Const {
         static let matchedCriteria = "itbl_matched_criteria"
         static let eventList = "itbl_event_list"
         static let anonymousUsageTrack = "itbl_anonymous_usage_track"
+        static let isNotificationsEnabled = "itbl_isNotificationsEnabled"
+        static let hasStoredNotificationSetting = "itbl_hasStoredNotificationSetting"
+
         static let attributionInfoExpiration = 24
     }
     
@@ -114,6 +117,11 @@ enum Const {
     enum HttpHeader {
         static let location = "Location"
         static let setCookie = "Set-Cookie"
+    }
+    
+    enum RemoteNotification {
+        static let aps = "aps"
+        static let contentAvailable = "content-available"
     }
 }
 
@@ -252,6 +260,9 @@ enum JsonKey {
         }
     }
 
+    static let mobileFrameworkInfo = "mobileFrameworkInfo"
+    
+    static let frameworkType = "frameworkType"
     
 //    embedded
     static let embeddedSessionId = "session"
@@ -325,6 +336,7 @@ enum JsonKey {
         static let packageName = "packageName"
         static let sdkVersion = "SDKVersion"
         static let content = "content"
+        static let jsonOnly = "jsonOnly"
     }
     
     enum Payload {