+++ swift-sdk/Internal/InternalIterableAPI.swift
@@ -256,7 +256,7 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
 
     // MARK: - API Request Calls
     
-    func register(token: Data,
+    func register(token: String,
                   onSuccess: OnSuccessHandler? = nil,
                   onFailure: OnFailureHandler? = nil) {
         
@@ -267,23 +267,26 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
             onFailure?(errorMessage, nil)
             return
         }
-        
-        if !isEitherUserIdOrEmailSet() && localStorage.userIdAnnon == nil {
+
+         if !isEitherUserIdOrEmailSet() && localStorage.userIdAnnon == nil {
             if config.enableAnonActivation {
                 anonymousUserManager.trackAnonTokenRegistration(token: token.hexString())
             }
             onFailure?("Iterable SDK must be initialized with an API key and user email/userId before calling SDK methods", nil)
             return
-        }
+        }       
+        hexToken = token
         
-        hexToken = token.hexString()
-        let registerTokenInfo = RegisterTokenInfo(hexToken: token.hexString(),
-                                                  appName: appName,
-                                                  pushServicePlatform: config.pushPlatform,
-                                                  apnsType: dependencyContainer.apnsTypeChecker.apnsType,
-                                                  deviceId: deviceId,
-                                                  deviceAttributes: deviceAttributes,
-                                                  sdkVersion: localStorage.sdkVersion)
+        let mobileFrameworkInfo = config.mobileFrameworkInfo ?? createDefaultMobileFrameworkInfo()
+        
+        let registerTokenInfo = RegisterTokenInfo(hexToken: token,
+                                                appName: appName,
+                                                pushServicePlatform: config.pushPlatform,
+                                                apnsType: dependencyContainer.apnsTypeChecker.apnsType,
+                                                deviceId: deviceId,
+                                                deviceAttributes: deviceAttributes,
+                                                sdkVersion: localStorage.sdkVersion,
+                                                mobileFrameworkInfo: mobileFrameworkInfo)
         requestHandler.register(registerTokenInfo: registerTokenInfo,
                                 notificationStateProvider: notificationStateProvider,
                                 onSuccess: { (_ data: [AnyHashable: Any]?) in
@@ -297,6 +300,12 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         )
     }
     
+    func register(token: Data,
+                  onSuccess: OnSuccessHandler? = nil,
+                  onFailure: OnFailureHandler? = nil) {
+        register(token: token.hexString(), onSuccess: onSuccess, onFailure: onFailure)
+    }
+    
     @discardableResult
     func disableDeviceForCurrentUser(withOnSuccess onSuccess: OnSuccessHandler? = nil,
                                      onFailure: OnFailureHandler? = nil) -> Pending<SendRequestValue, SendRequestError> {
@@ -305,11 +314,17 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
             onFailure?(errorMessage, nil)
             return SendRequestError.createErroredFuture(reason: errorMessage)
         }
+        
         guard userId != nil || email != nil else {
             let errorMessage = "either userId or email must be present"
             onFailure?(errorMessage, nil)
             return SendRequestError.createErroredFuture(reason: errorMessage)
         }
+        
+        // We need to call register token here so that we can trigger the device registration
+        // with the updated notification settings
+        
+        register(token: hexToken)
         
         return requestHandler.disableDeviceForCurrentUser(hexToken: hexToken, withOnSuccess: onSuccess, onFailure: onFailure)
     }
@@ -655,6 +670,8 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
     private var _userId: String?
     private var _successCallback: OnSuccessHandler? = nil
     private var _failureCallback: OnFailureHandler? = nil
+    
+    private let notificationCenter: NotificationCenterProtocol
 
     
     /// the hex representation of this device token
@@ -839,6 +856,7 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         //localStorage.email = nil      // remove this before pushing the code (only for testing)
         inAppDisplayer = dependencyContainer.inAppDisplayer
         urlOpener = dependencyContainer.urlOpener
+        notificationCenter = dependencyContainer.notificationCenter
         deepLinkManager = DeepLinkManager(redirectNetworkSessionProvider: dependencyContainer)
     }
     
@@ -871,15 +889,58 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         requestHandler.start()
         
         checkRemoteConfiguration()
+        
+        addForegroundObservers()
                 
         return inAppManager.start()
+    }
+    
+    private func addForegroundObservers() {
+        notificationCenter.addObserver(self,
+                                     selector: #selector(onAppDidBecomeActiveNotification(notification:)),
+                                     name: UIApplication.didBecomeActiveNotification,
+                                     object: nil)
+    }
+    
+    @objc private func onAppDidBecomeActiveNotification(notification: Notification) {
+        guard config.autoPushRegistration else { return }
+        
+        notificationStateProvider.isNotificationsEnabled { [weak self] systemEnabled in
+            guard let self = self else { return }
+            
+            let storedEnabled = self.localStorage.isNotificationsEnabled
+            let hasStoredPermission = self.localStorage.hasStoredNotificationSetting
+            
+            if self.isEitherUserIdOrEmailSet() {
+                if hasStoredPermission && (storedEnabled != systemEnabled) {
+                    if !systemEnabled {
+                        self.disableDeviceForCurrentUser()
+                    } else {
+                        self.notificationStateProvider.registerForRemoteNotifications()
+                    }
+                }
+                
+                // Always store the current state
+                self.localStorage.isNotificationsEnabled = systemEnabled
+                self.localStorage.hasStoredNotificationSetting = true
+            }
+        }
     }
     
     private func handle(launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
         guard let launchOptions = launchOptions else {
             return
         }
+        
         if let remoteNotificationPayload = launchOptions[UIApplication.LaunchOptionsKey.remoteNotification] as? [AnyHashable: Any] {
+            
+            if let aps = remoteNotificationPayload[Const.RemoteNotification.aps] as? [String: Any],
+               let contentAvailable = aps[Const.RemoteNotification.contentAvailable] as? Int,
+               contentAvailable == 1 {
+                ITBInfo("Received push notification with wakey content-available flag")
+                return
+            }
+            
             if let _ = IterableUtil.rootViewController {
                 // we are ready
                 IterableAppIntegration.implementation?.performDefaultNotificationAction(remoteNotificationPayload)
@@ -943,10 +1004,22 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
                 print("Error converting dictionary to data: \(error)")
             }
         }
+	}
+
+    private func createDefaultMobileFrameworkInfo() -> IterableAPIMobileFrameworkInfo {
+        let frameworkType = IterableAPIMobileFrameworkDetector.frameworkType()
+        return IterableAPIMobileFrameworkInfo(
+            frameworkType: frameworkType,
+            iterableSdkVersion: frameworkType == .native ? localStorage.sdkVersion : nil
+        )
     }
     
     deinit {
         ITBInfo()
+        notificationCenter.removeObserver(self)
         requestHandler.stop()
     }
+    
 }
+
+