+++ swift-sdk/Internal/Utilities/LocalStorageProtocol.swift
@@ -31,6 +31,10 @@ protocol LocalStorageProtocol {
 
     var anonymousSessions: IterableAnonSessionsWrapper? { get set }
     
+    var isNotificationsEnabled: Bool { get set }
+    
+    var hasStoredNotificationSetting: Bool { get set }
+    
     func getAttributionInfo(currentDate: Date) -> IterableAttributionInfo?
     
     func save(attributionInfo: IterableAttributionInfo?, withExpiration expiration: Date?)