+++ swift-sdk/Internal/in-app/InAppManager+Functions.swift
@@ -6,7 +6,7 @@ import Foundation
 
 enum MessagesProcessorResult {
     case show(message: IterableInAppMessage, messagesMap: OrderedDictionary<String, IterableInAppMessage>)
-    case noShow(messagesMap: OrderedDictionary<String, IterableInAppMessage>)
+    case noShow(message: IterableInAppMessage?, messagesMap: OrderedDictionary<String, IterableInAppMessage>)
 }
 
 struct MessagesProcessor {
@@ -30,14 +30,18 @@ struct MessagesProcessor {
         case let .skip(message):
             updateMessage(message, didProcessTrigger: true)
             return processMessages()
+        case let .skipAndConsume(message):
+            updateMessage(message, didProcessTrigger: true, consumed: true)
+            return .noShow(message: message, messagesMap: messagesMap)
         case .none, .wait:
-            return .noShow(messagesMap: messagesMap)
+            return .noShow(message: nil, messagesMap: messagesMap)
         }
     }
     
     private enum ProcessNextMessageResult {
         case show(IterableInAppMessage)
         case skip(IterableInAppMessage)
+        case skipAndConsume(IterableInAppMessage)
         case none
         case wait
     }
@@ -59,7 +63,11 @@ struct MessagesProcessor {
         
         ITBDebug("isOkToShowNow")
         
-        if inAppDelegate.onNew(message: message) == .show {
+        let returnValue = inAppDelegate.onNew(message: message)
+        if message.isJsonOnly {
+            return .skipAndConsume(message)
+        }
+        if returnValue == .show {
             ITBDebug("delegate returned show")
             return .show(message)
         } else {