+++ tests/common/MockLocalStorage.swift
@@ -34,6 +34,10 @@ class MockLocalStorage: LocalStorageProtocol {
     
     var anonymousUserUpdate: [AnyHashable : Any]?
 
+    var isNotificationsEnabled: Bool = false
+    
+    var hasStoredNotificationSetting: Bool = false
+    
     func getAttributionInfo(currentDate: Date) -> IterableAttributionInfo? {
         guard !MockLocalStorage.isExpired(expiration: attributionInfoExpiration, currentDate: currentDate) else {
             return nil