+++ tests/unit-tests/IterableAPITests.swift
@@ -495,6 +495,11 @@ class IterableAPITests: XCTestCase {
             TestUtils.validateMatch(keyPath: KeyPath(string: "device.dataFields.reactNativeSDKVersion"), value: "x.xx.xxx", inDictionary: body)
             TestUtils.validateNil(keyPath: KeyPath(string: "device.dataFields.\(attributeToAddAndRemove)"), inDictionary: body)
             
+            TestUtils.validateMatch(keyPath: KeyPath(string: "device.dataFields.mobileFrameworkInfo.frameworkType"), value: "native", inDictionary: body)
+
+            
+            TestUtils.validateMatch(keyPath: KeyPath(string: "device.dataFields.mobileFrameworkInfo.iterableSdkVersion"), value:  IterableAPI.sdkVersion, inDictionary: body)
+            
             expectation.fulfill()
         }) { reason, _ in
             // failure
@@ -1310,4 +1315,5 @@ class IterableAPITests: XCTestCase {
         XCTAssertEqual(localStorage.authToken, authToken)
         userDefaults.removePersistentDomain(forName: "upgrade.test")
     }
+
 }